requirements.txt
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backcall==0.2.0
bcrypt==3.2.2
boto3==1.24.59
botocore==1.27.59
CacheControl==0.12.11
cachy==0.3.0
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
cleo==0.8.1
clikit==0.6.2
colorama==0.4.5
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.3.1
cryptography==37.0.4
daphne==3.0.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-extensions==3.2.0
django-ninja==0.19.1
djangorestframework==3.13.1
dnspython==2.2.1
drf-yasg==1.21.3
email-validator==1.2.1
executing==0.10.0
filelock==3.6.0
gunicorn==20.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
inflection==0.5.1
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
keyring==23.5.0
lockfile==0.12.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
msgpack==1.0.3
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.8.2
platformdirs==2.5.2
poetry==1.1.13
poetry-core==1.0.8
prompt-toolkit==3.0.30
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.2
Pygments==2.13.0
PyJWT==2.4.0
pylev==1.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
pywin32-ctypes==0.2.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
service-identity==21.1.0
shellingham==1.4.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
tomlkit==0.10.2
traitlets==5.3.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
wincertstore==0.2
zipp==3.8.0
zope.interface==5.4.0
